fix: LaunchedEffect 동작 수정 - 조건문 추가

+++ AOS/app/src/main/java/com/example/eeos/presentation/topappbar/EeosTopAppBar.kt
@@ -40,7 +40,9 @@ fun EeosTopAppBar(
     }
 ) {
     LaunchedEffect(topAppBarUiState.value.isAccountDeleted) {
-        onDeleteSuccess()
+        if (topAppBarUiState.value.isAccountDeleted) {
+            onDeleteSuccess()
+        }
     }
 
     val deleteAccountDialogState = remember { mutableStateOf(false) }
